move constants file and update quiz comments
