+++ frontend/src/api/axiosInstance.ts
@@ -2,7 +2,6 @@ import axios from 'axios';
 
 // Use NEXT_PUBLIC prefix for client-side environment variables in Next.js
 const apiUrl = process.env.NEXT_PUBLIC_API_URL || 'http://127.0.0.1:5000';
-console.log('API URL being used:', apiUrl);
 
 const axiosInstance = axios.create({
 	baseURL: apiUrl,
@@ -11,6 +10,7 @@ const axiosInstance = axios.create({
 // this request interceptor adds the auth header to all requests
 axiosInstance.interceptors.request.use(
 	(config) => {
+		// get the auth token from localStorage
 		const token = localStorage.getItem('access_token');
 		if (token) {
 			config.headers.Authorization = `Bearer ${token}`;
